src/infrastructure/api/routes/urls.py
from django.conf import settings
from django.conf.urls import include
from django.urls import path

from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView


urlpatterns = [
    path('', include(f'{settings.API_ROUTES}.post.urls')),
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]
