pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Blank <rojermen96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
django-taggit = "^5.0.1"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg = "^3.1.19"
marshmallow = "^3.21.3"
djangorestframework-simplejwt = "^5.3.1"
pytest = "^8.2.2"
pre-commit = "^4.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
exclude_dirs = ["tests", "migrations", ".venv"]
tests = ["B101", "B102", "B110", "B404", "B509"]

[tool.isort]
profile = "black"
line_length = 90
known_first_party = ["api", "config", "utils", "services"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = ["*/migrations/*"]

[tool.black]
line-length = 90
skip-string-normalization = true
skip-magic-trailing-comma = false
exclude = '''
/(
  migrations
)/
'''
